:recycle: refactor: Refatoracao do método de login e autenticacao de usuarios
